move common constant to home modules
